requirements.txt
annotated-types==0.7.0
anyio==4.9.0
build==1.2.2.post1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cleo==2.1.0
crashtest==0.4.1
cryptography==44.0.3
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
dulwich==0.22.8
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.97.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.20.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
jiter==0.10.0
keyring==25.6.0
more-itertools==10.7.0
msgpack==1.1.0
numpy==2.3.0
openai==1.95.1
packaging==25.0
pbs-installer==2025.4.9
pexpect==4.9.0
pkginfo==1.12.1.2
platformdirs==4.3.8
pluggy==1.6.0
poetry==2.1.1
poetry-core==2.1.1
poetry-plugin-shell==1.0.1
proto-plus==1.26.1
protobuf==6.31.1
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.6
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyproject_hooks==1.2.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytube==15.0.0
RapidFuzz==3.13.0
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
SecretStorage==3.3.3
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
tomlkit==0.13.2
tqdm==4.67.1
trove-classifiers==2025.5.9.12
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
virtualenv==20.31.2
websockets==15.0.1
yt-dlp==2025.6.30
zstandard==0.23.0
google-cloud-texttospeech
google-cloud-speech
